move exception handler to middleware
